+++ node/src/test/kotlin/net/corda/node/services/vault/VaultQueryTests.kt
@@ -4,6 +4,8 @@ import net.corda.core.contracts.*
 import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.generateKeyPair
 import net.corda.core.crypto.toStringShort
+import net.corda.core.crypto.*
+import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.internal.packageName
 import net.corda.core.node.services.*
@@ -13,6 +15,9 @@ import net.corda.core.utilities.NonEmptySet
 import net.corda.core.utilities.days
 import net.corda.core.utilities.seconds
 import net.corda.core.utilities.toHexString
+import net.corda.core.transactions.LedgerTransaction
+import net.corda.core.transactions.TransactionBuilder
+import net.corda.core.utilities.*
 import net.corda.finance.*
 import net.corda.finance.contracts.CommercialPaper
 import net.corda.finance.contracts.Commodity
@@ -28,6 +33,7 @@ import net.corda.nodeapi.internal.persistence.CordaPersistence
 import net.corda.nodeapi.internal.persistence.DatabaseConfig
 import net.corda.testing.core.*
 import net.corda.testing.internal.TEST_TX_TIME
+import net.corda.testing.internal.chooseIdentity
 import net.corda.testing.internal.rigorousMock
 import net.corda.testing.internal.vault.DUMMY_LINEAR_CONTRACT_PROGRAM_ID
 import net.corda.testing.internal.vault.DummyLinearContract
@@ -98,16 +104,20 @@ class VaultQueryTests {
             "net.corda.testing.contracts",
             "net.corda.finance.contracts",
             CashSchemaV1::class.packageName,
-            DummyLinearStateSchemaV1::class.packageName)
-    private lateinit var services: MockServices
-    private lateinit var vaultFiller: VaultFiller
-    private lateinit var vaultFillerCashNotary: VaultFiller
-    private lateinit var notaryServices: MockServices
-    private val vaultService: VaultService get() = services.vaultService
-    private lateinit var identitySvc: IdentityService
-    private lateinit var database: CordaPersistence
-    @Before
-    fun setUp() {
+            DummyLinearStateSchemaV1::class.packageName,
+            SampleCashSchemaV3::class.packageName,
+            VaultQueryTestsBase.MyContractClass::class.packageName)
+
+    override lateinit var services: MockServices
+    override lateinit var vaultFiller: VaultFiller
+    override lateinit var vaultFillerCashNotary: VaultFiller
+    override lateinit var notaryServices: MockServices
+    override val vaultService: VaultService get() = services.vaultService
+    override lateinit var identitySvc: IdentityService
+    override lateinit var database: CordaPersistence
+
+
+    override fun before() {
         // register additional identities
         val databaseAndServices = makeTestDatabaseAndMockServices(
                 cordappPackages,
@@ -208,6 +218,43 @@ class VaultQueryTests {
             vaultService.queryBy<FungibleAsset<*>>(criteria)
         }
     }
+
+    @Test
+    fun `query by interface for a contract class extending a parent contract class`() {
+        database.transaction {
+
+            // build custom contract and store in vault
+            val me = services.myInfo.chooseIdentity()
+            val state = MyState("myState", listOf(me))
+            val stateAndContract = StateAndContract(state, MYCONTRACT_ID)
+            val utx = TransactionBuilder(notary = notaryServices.myInfo.singleIdentity()).withItems(stateAndContract).withItems(dummyCommand())
+            services.recordTransactions(services.signInitialTransaction(utx))
+
+            // query vault by Child class
+            val criteria = VaultQueryCriteria() // default is UNCONSUMED
+            val queryByMyState = vaultService.queryBy<MyState>(criteria)
+            assertThat(queryByMyState.states).hasSize(1)
+
+            // query vault by Parent class
+            val queryByBaseState = vaultService.queryBy<BaseState>(criteria)
+            assertThat(queryByBaseState.states).hasSize(1)
+
+            // query vault by extended Contract Interface
+            val queryByContract = vaultService.queryBy<MyContractInterface>(criteria)
+            assertThat(queryByContract.states).hasSize(1)
+        }
+    }
+
+    // Beware: do not use `MyContractClass::class.qualifiedName` as this returns a fully qualified name using "dot" notation for enclosed class
+    val MYCONTRACT_ID = "net.corda.node.services.vault.VaultQueryTestsBase\$MyContractClass"
+
+    open class MyContractClass : Contract {
+        override fun verify(tx: LedgerTransaction) {}
+    }
+
+    interface MyContractInterface : ContractState
+    open class BaseState(override val participants: List<AbstractParty> = emptyList()) : MyContractInterface
+    data class MyState(val name: String, override val participants: List<AbstractParty> = emptyList()) : BaseState(participants)
 
     @Test
     fun `unconsumed states simple`() {